+++ tinyc/tests/pp/04.c
@@ -1,4 +1,0 @@
-#define foobar 1
-#define C(x,y) x##y
-#define D(x) (C(x,bar))
-D(foo)