+++ changelog.md
@@ -10,10 +10,11 @@
 
 ## Compiler changes
 
+- Specific warnings can now be turned into errors via `--warningAsError[X]:on|off`.
+- The `define` and `undef` pragmas have been de-deprecated.
 - JavaScript backend adds compile-time optional `let` variable declarations,
-  compile Nim with `-d:jsLet` for `let` and without for `var`,
+  compile Nim with `-d:nimJsVar` for `var` and without for `let`,
   for a transition period, for more information see https://caniuse.com/let
-
 
 ## Tool changes